updated repo to estnet v1.0
This commit updates the Github repository to the latest version of ESTNeT

+++ src/estnet/node/tracking/swipe/ExponentialSwipeTracking.h
@@ -49,7 +49,7 @@ protected:
      */
     virtual double applyScaling(double ratio);
 
-    int _exponent;
+    double _exponent;
 };
 
 } //namespace
